Render each comment dynamically.
The complexity of the application took a big jump in this example. Let's
take the changes step by step.

1. We added a top level array of objects stored in the variable data.
Each object contains an id, author, and text key.

2. We modified our instantiation of CommentBox so that the data array is
available to it in this.props.data. In other words, we *instantiated* a
CommentBox, with its data attribute set to the data array. This is an
important step.  Data can only flow from the parent component to a child
component, so CommentBox, being the top-most React Component, needs to
have the data for it to be available to any child component under
CommentBox.

3. We modified the CommentBox class itself so that it passes on the data
that it holds in this.props.data to the CommentList component. Again,
this is done by setting the data attribute of the CommentList to
this.props.data.

4. We then modified CommentList so that it maps over its data (held in
this.props.data), and turns each one of those objects into a React
Comment component. Each time it maps over an object in this.props.data,
it creates a Comment, and passes in the data to that comment in the form
of attributes that the Comment component knows how to use. Note that, up
until this point, we have kept the data in one package (i.e, as data).
Finally, and very interestingly, it returns an **array** of these React
Comment Components, which React must *implicitly* iterate over and treat
as nested components.

The important thing to remember is that data flows one way, from parent
to child. Data is passed in a React Component's attributes, and become
available to the component via this.props. Text becomes available to the
component in this.props.children, although this can contain other data.
Finally, when a component returns an array of components, React must
implicitly iterate over them.