Change CommentBox to dynamically fetch data from the server.
Defined `getInitialState` in the CommentBox component to *set* the
initial state of the coomponent. This state is passed down to the
CommentList component in the CommentList's `data` property.

Defined `componentDidMount`, which is a function that React will
automatically run when a component is rendered for the first time. It
was in the `componentDidMount` function that we define an AJAX call to
the server. It is important to note that **the key to dynamically
rendering data** is that our `componentDidMount` function changes the
CommenBox's state. This change in state propagates down to child /
nested components.