Update CommentForm component so that it is able to submit new comments to the server.
There is a lot going on this commit. Let's go over them one by one.

1. First, we have added an `getInitialState` function to our component
in order to set the initial state of the component. Again, the name of
the function is odd, since we use it to set the initial state. Remember
that this will only be run once in the lifetime of a component.

2. We added two functions - `handleAuthorChange`, `handleTextChange` -
that make sure that when our user types into one of the input elements,
that input is reflected in our component. We reflect it in the component
by updating the component's state with `setState`. Notice that we are
setting the `author` or `text` property of the `state` object - which
lives on the component - to a particular value.

3. We added a `handleSubmit` function whose job it is to send the data
to the `CommentBox` component and to change reset the form by changing
the state of the `CommentForm` component. The first part deserves some
explanation. The only way that we are able to pass data from the child
to the parent is by passing in a callback function that lives on the
parent as a property of the child. In other words, the function
`handleCommentSubmit` is defined on the CommentBox component. It is
passed down as a property on the CommentForm component. By asking
CommentForm to invoke this callback, we can effectively pass data to the
parent CommentBox.